requirements.txt
Flask==3.0.3
flask-cors==4.0.1
openai==1.51.2
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
chromadb==0.5.11
pypdf==5.0.1
python-docx==1.1.2
python-pptx==1.0.2
tiktoken==0.8.0
